Добавил  ProductController, обновил ProductDAO

+++ src/main/java/com/itechart/stockOnline/dao/ProductDao.java
@@ -4,5 +4,5 @@ import com.itechart.stockOnline.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ProductDao extends JpaRepository<Product, Integer>, JpaSpecificationExecutor {
+public interface ProductDao extends JpaRepository<Product, Long>, JpaSpecificationExecutor {
 }
